+++ test/ext/nil_test.exs
@@ -1,0 +1,5 @@
+defmodule Ext.NilTest do
+  use ExUnit.Case, async: true
+
+  doctest Ext.Nil
+end