+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added `Functor.map/2`.
+- Added `Nil.map/2`.
 
 ## 1.3.0